test_app/requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver
python-dotenv>=0.19.0   # For environment variables